Updated parser to remove posts made private
The parser now keeps of all posts that are currently on
piazza and has a status other than 'deleted' or 'private'.
It then checks if any of the posts in the databse is not
included in the colletion of posts currently on piazza.
If so, that post is deleted from the database.

This is done in order to make sure that posts that were
once public but have been changed to private are removed
from the database.

Unfortunately it's not enough to check if the posts have
status 'private', as some private posts aren't returned
when fetching posts from piazza, while others are.